requirements.txt
# Запуск: pip install -r requirements.txt
customtkinter>=5.2
Pillow>=10.0
tkcalendar>=1.6
